requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.4
# django-admin==2.0.2  # Comentado: depende de paquetes problemáticos en Windows y no se usa en el proyecto
django-allauth==0.55.0
django-cors-headers==4.2.0
django-environ==0.10.0
# django-excel-response2==3.0.6  # Comentado: puede tener dependencias problemáticas en Windows y no se usa
django-jazzmin==3.0.1
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
# excel-base==1.0.4  # Comentado: su dependencia 'screen' requiere C++ Build Tools en Windows y no se usa
fonttools==4.49.0
freezegun==1.2.2
gunicorn==21.2.0
html5lib==1.1
idna==3.4
isoweek==1.3.3
kombu==5.3.6
mailersend==0.5.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
# screen==1.0.1  # Comentado: requiere compilación C++ en Windows y no se usa en el proyecto
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
TimeConvert==3.0.13
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.0.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.1
webencodings==0.5.1
whitenoise==6.5.0
xlwt==1.3.0
zopfli==0.2.3
